refactor: move config file stuff from swingset-runner into vat controller
